Initial migration to resource conditions
