+++ fabric-resource-conditions-api-v1/build.gradle
@@ -1,3 +1,7 @@
 version = getSubprojectVersion(project)
 
-testDependencies(project, [':fabric-gametest-api-v1'])
+testDependencies(project, [
+	':fabric-gametest-api-v1',
+	':fabric-lifecycle-events-v1',
+	':fabric-resource-loader-v0'
+])
