+++ fabric-resource-conditions-api-v1/src/main/java/net/fabricmc/fabric/api/resource/conditions/v1/ResourceConditions.java
@@ -85,6 +85,17 @@ public final class ResourceConditions {
 	 * The key ({@value}) Fabric uses to identify resource conditions in a JSON object.
 	 */
 	public static final String CONDITIONS_KEY = "fabric:load_conditions";
+
+	/**
+	 * The key ({@value}) Fabric uses to identify if custom overlays are in a pack.mcmeta file.
+	 */
+	public static final String OVERLAYS_KEY = "fabric:overlays";
+
+	/**
+	 * The key ({@value}) Fabric uses to identify custom overlays in a custom overlay object.
+	 */
+	public static final String ENTRIES_KEY = "entries";
+
 	/**
 	 * The key ({@value}) identifying the resource condition's identifier inside a condition object.
 	 */
@@ -166,7 +177,7 @@ public final class ResourceConditions {
 		Predicate<JsonObject> jrc = get(conditionId);
 
 		if (jrc == null) {
-			throw new JsonParseException("Unknown recipe condition: " + conditionId);
+			throw new JsonParseException("Unknown resource condition: " + conditionId);
 		} else {
 			return jrc.test(condition);
 		}
