+++ fabric-resource-conditions-api-v1/src/main/resources/fabric-resource-conditions-api-v1.mixins.json
@@ -7,6 +7,7 @@
     "DataProviderMixin",
     "JsonDataLoaderMixin",
     "LootManagerMixin",
+    "ResourcePackProfileMixin",
     "SinglePreparationResourceReloaderMixin",
     "TagManagerLoaderMixin"
   ],
