+++ fabric-resource-conditions-api-v1/src/testmod/resources/fabric.mod.json
@@ -9,6 +9,9 @@
     "fabric-resource-conditions-api-v1": "*"
   },
   "entrypoints": {
+    "main": [
+      "net.fabricmc.fabric.test.resource.conditions.OverlayConditionsTest"
+    ],
     "fabric-gametest": [
       "net.fabricmc.fabric.test.resource.conditions.ConditionalResourcesTest"
     ]
