+++ fabric-resource-conditions-api-v1/src/testmod/resources/resourcepacks/overlay_test/do_overlay/data/fabric-resource-conditions-api-v1-testmod/recipes/do_overlay.json
