+++ fabric-resource-conditions-api-v1/src/testmod/resources/resourcepacks/overlay_test/dont_overlay/data/fabric-resource-conditions-api-v1-testmod/recipes/dont_overlay.json
