+++ fabric-resource-conditions-api-v1/src/testmod/resources/resourcepacks/overlay_test/pack.mcmeta
@@ -1,0 +1,27 @@
+{
+  "pack": {
+    "pack_format": 26,
+    "description": "Fabric Overlay Conditions Test Datapack."
+  },
+  "fabric:overlays": {
+    "entries": [
+      {
+        "directory": "do_overlay",
+        "condition": "fabric:all_mods_loaded",
+        "values": [
+          "fabric-resource-conditions-api-v1"
+        ]
+      },
+      {
+        "directory": "dont_overlay",
+        "condition": "fabric:not",
+        "value": {
+          "condition": "fabric:all_mods_loaded",
+          "values": [
+            "fabric-resource-conditions-api-v1"
+          ]
+        }
+      }
+    ]
+  }
+}
