+++ fabric-resource-loader-v0/src/main/java/net/fabricmc/fabric/impl/resource/loader/ResourceManagerHelperImpl.java
@@ -128,7 +128,6 @@ public class ResourceManagerHelperImpl implements ResourceManagerHelper {
 							return pack;
 						} else {
 							List<ResourcePack> overlays = new ArrayList<>(metadata.overlays().size());
-
 							for (String overlay : metadata.overlays()) {
 								overlays.add(pack.createOverlay(overlay));
 							}
