+++ fabric-resource-loader-v0/src/main/java/net/fabricmc/fabric/mixin/resource/loader/ResourcePackProfileMixin.java
@@ -16,15 +16,8 @@
 
 package net.fabricmc.fabric.mixin.resource.loader;
 
-import java.io.IOException;
-import java.util.ArrayList;
-import java.util.List;
 import java.util.Set;
 import java.util.function.Predicate;
-
-import com.llamalad7.mixinextras.sugar.Local;
-
-import net.fabricmc.fabric.impl.resource.loader.OverlayConditionsMetadata;
 
 import org.spongepowered.asm.mixin.Final;
 import org.spongepowered.asm.mixin.Mixin;
@@ -32,7 +25,6 @@ import org.spongepowered.asm.mixin.Shadow;
 import org.spongepowered.asm.mixin.Unique;
 import org.spongepowered.asm.mixin.injection.At;
 import org.spongepowered.asm.mixin.injection.Inject;
-import org.spongepowered.asm.mixin.injection.ModifyVariable;
 import org.spongepowered.asm.mixin.injection.callback.CallbackInfoReturnable;
 
 import net.minecraft.resource.ResourcePack;
@@ -79,15 +71,4 @@ abstract class ResourcePackProfileMixin implements FabricResourcePackProfile {
 		this.parentsPredicate = predicate;
 	}
 
-	@ModifyVariable(method = "loadMetadata", at = @At("STORE"))
-	private static List<String> fabric_applyOverlayConditions(List<String> overlays, @Local ResourcePack resourcePack) throws IOException {
-		List<String> appliedOverlays = new ArrayList<>(overlays);
-		OverlayConditionsMetadata overlayConditionsMetadata = resourcePack.parseMetadata(OverlayConditionsMetadata.SERIALIZER);
-
-		if (overlayConditionsMetadata != null) {
-			appliedOverlays.addAll(overlayConditionsMetadata.getAppliedOverlays());
-		}
-
-		return appliedOverlays;
-	}
 }
