+++ fabric-resource-loader-v0/src/testmod/resources/fabric.mod.json
@@ -12,8 +12,7 @@
     "main": [
       "net.fabricmc.fabric.test.resource.loader.BuiltinResourcePackTestMod",
       "net.fabricmc.fabric.test.resource.loader.ResourceReloadListenerTestMod",
-      "net.fabricmc.fabric.test.resource.loader.VanillaBuiltinResourcePackInjectionTestMod",
-      "net.fabricmc.fabric.test.resource.loader.OverlayConditionsTestMod"
+      "net.fabricmc.fabric.test.resource.loader.VanillaBuiltinResourcePackInjectionTestMod"
     ],
     "server": [
       "net.fabricmc.fabric.test.resource.loader.LanguageTestMod"
