--- fabric-resource-loader-v0/src/testmod/resources/resourcepacks/overlay_test/pack.mcmeta
@@ -1,23 +1,0 @@
-{
-  "pack": {
-    "pack_format": 26,
-    "description": "Fabric Overlay Conditions Test Datapack."
-  },
-  "fabric:overlays": {
-    "entries": [
-      {
-        "directory": "do_overlay",
-        "type": "fabric:mod_loaded",
-        "mod_id": "fabric-resource-loader-v0-testmod"
-      },
-      {
-        "directory": "dont_overlay",
-        "type": "fabric:not",
-        "condition": {
-          "type": "fabric:mod_loaded",
-          "mod_id": "fabric-resource-loader-v0-testmod"
-        }
-      }
-    ]
-  }
-}
